SwapChainVk: Add frame pacing by waiting on previous rendering on images
Vulkan doesn't guarantee any frame pacing on its own even if a number of
drivers do it internally. This means that many rendering + present
can be in flight at the same time, making the mouse-to-screen latency
terrible.

This adds frame pacing in a simple way by waiting on previous rendering
on textures to be finished before returning it in GetCurrentTexture.

Bug: 42241496
Change-Id: I01d3a1b69cc8f86f01c384c0a1d38400d4f01790